Role based authentication on delete controllers

+++ src/main/java/com/sunil/blog/controllers/UserController.java
@@ -2,11 +2,14 @@ package com.sunil.blog.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,10 +19,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
+import com.sunil.blog.entities.User;
 import com.sunil.blog.payloads.ApiResponse;
 import com.sunil.blog.payloads.UserDto;
 import com.sunil.blog.service.UserService;
+import com.sunil.blog.utils.RequestHelper;
 
 @RestController
 @CrossOrigin
@@ -28,6 +32,9 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private RequestHelper requestHelper;
 
     @PostMapping("/add")
     public ResponseEntity<UserDto> createUser(@Valid @RequestBody UserDto userDto) {
@@ -42,11 +49,29 @@ public class UserController {
         return new ResponseEntity<>(updatedUser, HttpStatus.OK);
     }
 
+    /// User can only delete to himself
+    ///
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<ApiResponse> deleteUser(@PathVariable Integer id) {
-        this.userService.deleteUser(id);
-        ApiResponse apiResponse = new ApiResponse("User Deleted Successfully", true);
+    public ResponseEntity<ApiResponse> deleteUser(@PathVariable Integer id, HttpServletRequest request) {
+        User currentUser = this.requestHelper.getUserFromRequest(request);
 
+        if (currentUser.getId() != id)
+            throw new AccessDeniedException("Access is denied for delete User");
+
+        this.userService.deleteUser(id);
+
+        ApiResponse apiResponse = new ApiResponse("User Deleted Successfully", true);
+        return new ResponseEntity<>(apiResponse, HttpStatus.OK);
+    }
+
+    /// Only admin can delete any user
+    ///
+    @PreAuthorize("hasRole('ADMIN')")
+    @DeleteMapping("/admin/delete/{id}")
+    public ResponseEntity<ApiResponse> deleteUserByAdmin(@PathVariable Integer id) {
+        this.userService.deleteUser(id);
+
+        ApiResponse apiResponse = new ApiResponse("User Deleted Successfully", true);
         return new ResponseEntity<>(apiResponse, HttpStatus.OK);
     }
